requirements.txt
honcho==1.1.0
certifi==2023.5.7
charset-normalizer==2.0.7
click==8.0.3
cssselect==1.1.0
Flask==2.0.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
lxml==4.9.2
MarkupSafe==2.0.1
pyquery==1.4.3
requests==2.31.0
requests-toolbelt==0.9.1
setuptools==68.0.0
urllib3==1.26.7
Werkzeug==2.0.2
wheel==0.40.0
PyJWT==2.3.0
python-dotenv==0.19.1
python-decouple==3.5
Flask-PyMongo==2.3.0
cryptography==41.0.1
dnspython==2.1.0
rq==1.10.0
rq_scheduler==0.11.0
redis==4.6.0
numpy==1.22.4
exponent_server_sdk==2.0.0
pyfcm==1.5.2
aiohttp==3.8.1